Reorganize highlighting code
